rename iframe loaders consistently
