+++ catalog/app/components/Preview/loaders/IFrame/LoaderBrowsable.tsx
